Add an address name table (for use by logSymbol()). (dm)
git-svn-id: svn://mielke.cc/main/brltty@8357 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/dynld.h
@@ -27,7 +27,7 @@ extern void *loadSharedObject (const char *path);
 extern void unloadSharedObject (void *object);
 
 extern int findSharedSymbol (void *object, const char *symbol, void *pointerAddress);
-extern const char *getSharedSymbolName (void *address, unsigned int *offset);
+extern const char *getSharedSymbolName (void *address, intptr_t *offset);
 
 #ifdef __cplusplus
 }
