+++ Programs/dynld_dlfcn.c
@@ -76,7 +76,7 @@ findSharedSymbol (void *object, const char *symbol, void *pointerAddress) {
 }
 
 const char *
-getSharedSymbolName (void *address, unsigned int *offset) {
+getSharedSymbolName (void *address, intptr_t *offset) {
 #ifdef __USE_GNU
   Dl_info info;
